+++ scripts/utility/utility.js
@@ -1,5 +1,8 @@
 import { scene } from "./scene.js";
 
 export let utility = {
-    "scene": scene
+    "scene": scene,
+    "getControlledActor": function _getControlledActor() {
+        return canvas?.tokens?.controlled[0]?.actor
+    }
 };